uploading json
Used a python program to loop through a csv version of the original dataset and grabed 500 movies/tv shows from the US. Then dumped it all into a json with some formatting. Also added a UUID based on a hash of the title, can change this later but this should work.